Got ceilometer plotting
Added code for parsing all the ceilometer parameters and added a plot
function based on the lidar stare plot function.

There is a bit of a fudge in there that I create a "fake" header that
matches the lidar one. Really I think the best thing would be a generic
header class. Also note that I haven't yet sorted checking the checksum,
but I need to as I haven't really put any other checking in, so this
will be the check that everything is as it should be.