Some changes to the InstrumentPlotter class and lots of unicode string work
This commit started out with the intention of splitting reading data
from plotting data in the InstrumentPlotter class which has been done.
However in doing so and applying it to the child CeilometerPlotter
class I ended up having to propagate sci::strings replacing std::strings
through the whole code structure. That makes this a bit of a messy
commit, but there's not a lot I can do about it.