Restructured how we set up time and position in AmfNcFile, set up lidar wind profile output and temporarily commented out ceilometer output.
Now we don't write the time data on construction of the AmfNcFile, we
do this with a separate call. In the constructor we create the variables
only. This is because writing the data ends define mode so it means we
can't add further variables.

Commented out the ceilometer code because it needs modifying to reflect
these changes.