moved responsibility for filtering friles from the FolderChangesLister class to the InstrumentProcessor class.
This permits a derived InstrumentProcessor to use arbitrarily
complex proceedures to filter the files. Including regexs on the
full path and opening the files to check their contents.

The previous system (filename wildcards passed to the FolderChangesLister)
did not permit discrimination between the cross and co polarised
lidar stare data as the diles appear identical other than the folder.