Added a settings file
This is a very big commit with a lot of things in. The main change is that
I have added xml files for settings. One is a database file which holds
information about al the various "entities" that could be put fed into the
software. E.g. it has multiple people that can be authors or PIs. Then there
is a settings file that references entities in the database or it can
have new entities that are not referenced in the database. The settings file
also has other items that are not entities, such as folder locations, options
for autorunning or generating quicklooks etc.