Moved parsing setup file and added progress messages
The setup call now occurs when processing starts. This was because
it can take a long time to parse data from moving platforms so it
was not apropriate to do it in the window constructor.
Added progress messages to this too again because it takes a long time
during which the window appears to hang.
Also added the ability to change the setup file from a menu item and
removed the now defunct setting of the input and output directories.