Added checks for failed reads.
I had a read of a lidar fille that seemed to get stuck in an infinite
loop never finding the end of a file. I think this occured because a
read failed (it was a network file so maybe the connection was briefly
lost or someone else accessed the file and locked it), which meant all
future reads failed and therefore the stream never found the end of
the file.

Now the read functions check the stream is good at various points during
reading and throw if it is not.