Restructured how we pass instrument data to the instrument
Now we put a processor in the setup file and the processor node has
an instrument and a calibrationInfo.

This makes the main program more generic (it just deals with pointers
to InstrumentProcessors) and it makes the setu file more flexibe
because we don't have to run every processor for a particular instrument.