Sorted inconsistencies with decibels
The decibel nc variable was a mess and some compile errors were generated during the change to having Phsicals with a value type. Now everything in the Decibel variable uses linear units up to the point of generating data for netcdf file - i.e. returning from flattenData and writing the min and max values.