Changed the regex for finding lidar files
The 18 in the filename is the digits after the - in the instrument serial
number. This can be 1-3 characters. Also in older files this value isn't
there at all. Replaced .._ in the regex (which means match two of any
character then an underscore) with .{0,4} (which means match any character
0-4 times).

This will match any serial number or an old style file with no serial number.