Added lidar depolarisation processor
In order to reuse code from the stare processor, this processor contains
a cross stare processor and a co stare processor. An extra function has
been added to the stare processor to grab the data in a format that
is ready to be written to netcdf - i.e. padded where needed and transposed
to have the dimension order correct. This function is then used to grab the
data by the depol processor.
Once grabbed - further checks and padding are performed. The cross stare
should be immediately after the co stare. It is checked that the two types
alternate in time. Where they do not alternate in time extra padding profiles
are added. THERE IS NO CHECK THAT THE CROSS STARE IS IMMEDIATELY AFTER THE
CO STARE, because it is not obvious what immediate means - it may change
depending upon the scan pattern being used.