Reinstated lidar depol processor
The lidar depol ratio processor stopped functioning after the
changes to meet the latest standard. This was because the sparate
code for dealing with stare processors was removed as the stare data
became the same format as the scan data.
The overall code for plotting copolarised lidar data has had a
function added that grabs the data in the correct grid format.
This is now called by all processors including the depol ratio
processor to grab the data.