Changed how we find min/max for a variable
Each netcdf variable reports its min/max value. Up to now we filtered
out nans and fillvalues from this min/max. But Barbara has told me
That we should filter out any values with a flag not set to 1.

To do this we now pass a flag into the AmfNcVariable constructor.
The flag can be any integral type or a multi-d vector of an integral
type providing it has the same number or fewer dimensions than the
variable data. Any values with flags not equal to 1 will be ignored.

Where flag has fewer dimensions than the variable, we will cascade
down the dimensions of both the flag and variable, unti we reach
the integral flag. We will then continue cascading down the variable
dimensions applying the integral flag value to all variable values
found.