Added some extra qc checks to the micro rain radar
These include flagging negative numbers and values that are
unphysically high.

The max value for rainfall rate I've allowed is 20 m s-1. This is
probably faster than any hail that we are likely to measure that
isn't going to smash the dish up.

The max concentration I went for is 1e6 m-3 mm-1. This is probably
excessively high, but I thought it was better to be too high
rather than too low.