Fixed the dimensionality in lidar depol processor
In AMF standard version 1.1.0 the dimensionality of stare data
changed from 2d to 3d with the addition of the (size 1) angle
dimension. However the dimensionality of the depol stare data
remained 2d. Because the depol processor uses the stare code
to format, there were all sorts of problems with the
dimensionality.

These are now all fixed. In anticipation of the fact that the
dimensionality of this file might change in the future to match
the backscatter stare file, we keep the 3d data most of the
way through, then flatten and reshape it as we output it to file.