Fixed a couple of bugs with depol/stare lidar processing
I was getting an error thrown about mismatching profiles and once
that was fixed I was getting an infinite loop.

The mismatching profiles error was an unassigned variable for the
focus distance - fixed in HplHeader.cpp

The infinite loop was when padding the lidar depol data to ensure
both co and cross polarization data had the same number of profiles.

Both issues now fixed.