completed iwv-lwp netcdf file from microwave radiometer.
There is an issue in that the "status" flag from the raw data is
not part f the AMF standard, yet this flag should be used to flag
out bad data according to the manual. This can result in all flags
in the netcdf file being good, but either bad data is presented
and the user doesn't know, or data is set to fillvalue, but the
flags all indicate the data is good. No idea what to do about this.