Got the MRR code back up and running with the multi-D GridData class
I never converted the MRR code to use the GridData class rather than
std::vector. The code was just commented out.

Got this all sorted now. Also found and fixed a couple of bugs
for the AOF v2.0 standard because this is the first time I've used
this version of the standard as moving platforms cannot use it.