Fixed infinite loop in FolderChangesLister
When reading the previously processed files file, a bad line could
cause an infinite loop.

I had a sitation here the file got corrupted and hence the line being
read in was less than 19 characters, but it set the bad bit of the
fstream, meaning the next line never got read and the line variable
never became empty. This meant the code just sat in an infinite loop.

The infinite loop is now fixed, however, the code cannot get past a
ccorruption. Meaning it only reads the file up to a corupt line.

This is better than an infinite loop, but needs addressing at some
point.