Added support for the Gallion wind profile lidar
The code works when using the advanced setup. It tries to identify
VAD, RHI and stare sequences and splits them into separate output
files. It does a relatively simple wind profile calculation for the
VAD profiles. This can probably be improved, but I needed to get it
done relatively quickly.